requirements.txt
laspy>=2.5.0
numpy>=1.26.0
pyyaml>=6.0.0
trimesh>=4.4.0
pytest>=8.0.0
